service/tcfdreport-service/requirements.prod.txt
# FastAPI & Web (최소한의 기본 패키지만)
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1

# DB & Async (기본)
asyncpg==0.29.0
httpx==0.27.0
sqlalchemy==2.0.23

# JWT 인증
PyJWT==2.8.0

# 문서 처리 (Word, PDF 생성용)
python-docx==1.1.0
weasyprint==61.2
# WeasyPrint 의존성 (정확한 버전 고정)
pydyf==0.10.0
tinycss2==1.2.1
cssselect2==0.7.0
cairocffi>=0.9.0
cffi>=1.1.0

# PDF 생성 대안 (Debian 13 호환)
# pdfkit==1.0.0  # wkhtmltopdf 의존성으로 인해 제거

# Utility (기본)
pydantic==2.7.4
pydantic-settings==2.3.4

# Logging (기본)
structlog==23.2.0
